Get rid of a try:..except: pass.
git-svn-id: https://svn.erp5.org/repos/neo/trunk@1607 71dcc9de-d417-0410-9af5-da40c76e7ee4

+++ neo/connection.py
@@ -255,12 +255,9 @@ class Connection(BaseConnection):
 
             # Remove idle events, if appropriate packets were received.
             for msg_id in (None, packet.getId()):
-                try:
-                    event = self.event_dict[msg_id]
-                    del self.event_dict[msg_id]
+                event = self.event_dict.pop(msg_id, None)
+                if event is not None:
                     self.em.removeIdleEvent(event)
-                except KeyError:
-                    pass
 
             try:
                 packet_type = packet.getType()
